backend/requirements.txt
flask
flask-cors
flask-jwt-extended
python-dotenv
werkzeug
pymongo
faiss-cpu
sentence-transformers
beautifulsoup4
requests